[BREAKING Router] Rename StorybooksRouter to ClientRouter
