+++ src/Artsy/Router/__stories__/Router.story.tsx
@@ -1,6 +1,5 @@
 import { storiesOf } from "@storybook/react"
-import { StorybooksRouter } from "Artsy/Router"
-import { PreloadLink } from "Artsy/Router"
+import { ClientRouter, PreloadLink } from "Artsy/Router/client"
 import { ContextConsumer } from "Artsy/SystemContext"
 import React from "react"
 import { graphql } from "react-relay"
@@ -77,7 +76,7 @@ const routes = [
 
 storiesOf("SSR Router/Example", module).add("Example Router App", () => {
   return (
-    <StorybooksRouter
+    <ClientRouter
       routes={routes}
       context={{
         mediator: x => x,
