+++ src/Artsy/Router/client.ts
@@ -1,3 +1,3 @@
 export * from "./index"
 export { buildClientApp } from "./buildClientApp"
-export { StorybooksRouter } from "./Components/StorybooksRouter"
+export { ClientRouter } from "./Components/ClientRouter"
